notifications for console and user in services and actions

+++ src/components/student/StudentCourseList.js
@@ -2,10 +2,9 @@ import React, { useEffect } from 'react'
 import { connect } from 'react-redux'
 import { getStudentCourses, deleteAppliedCourse } from '../../reducers/actionCreators/studentActions'
 import CourseWithDel from './CourseWithDel'
-import { notify } from '../../reducers/actionCreators/notificationActions'
 import { Table } from 'react-bootstrap'
 
-export const StudentCourseList = ({ notify, loggedUser, courses, id, getStudentCourses, deleteAppliedCourse }) => {
+export const StudentCourseList = ({ loggedUser, courses, id, getStudentCourses, deleteAppliedCourse }) => {
 
   useEffect(() => {
     getStudentCourses(id)
@@ -16,8 +15,9 @@ export const StudentCourseList = ({ notify, loggedUser, courses, id, getStudentC
   //event handler for deleting specific course application, tells studentactions to deleteApliedCourse
   const removeApply = (id) => {
     return () => {
-      deleteAppliedCourse(id, loggedUser.user.user_id)
-      notify('Course is removed', 5)
+      if (window.confirm('Do you want to delete this course from your application?')) {
+        deleteAppliedCourse(id, loggedUser.user.user_id)
+      }
     }
   }
 
@@ -55,5 +55,5 @@ const mapStateToProps = (state) => {
 
 export default connect(
   mapStateToProps,
-  { getStudentCourses, deleteAppliedCourse, notify }
+  { getStudentCourses, deleteAppliedCourse }
 )(StudentCourseList)
